test: create TypeTests folder
