+++ tests/units/TypeTests/FloatTests.cs
@@ -3,14 +3,14 @@ using System.Security.Cryptography;
 using FingerprintBuilder.Tests.Models;
 using Xunit;
 
-namespace FingerprintBuilder.Tests;
+namespace FingerprintBuilder.Tests.TypeTests;
 
-public class DoubleTests
+public class FloatTests
 {
     private readonly Func<ThisUser, byte[]> _sha1;
     private readonly ThisUser _user;
 
-    public DoubleTests()
+    public FloatTests()
     {
         _sha1 = FingerprintBuilder<ThisUser>
             .Create(SHA1.Create())
@@ -18,7 +18,7 @@ public class DoubleTests
             .For(p => p.Number)
             .Build();
 
-        _user = new ThisUser { FirstName = "John", Number = 2.1d };
+        _user = new ThisUser { FirstName = "John", Number = 2.1f };
     }
 
     [Fact]
@@ -26,14 +26,14 @@ public class DoubleTests
     {
         var hash = _sha1(_user).ToLowerHexString();
 
-        Assert.Equal("af167d0996599a29b89ce1c467013bb7c98e7dcb", hash);
+        Assert.Equal("487b230857d43f392b9d20189d972d0cc8aa7c98", hash);
     }
 
     [Fact]
     public void UserInfo_Sha1_UpdateBool_ChangeHash()
     {
         var hash0 = _sha1(_user).ToLowerHexString();
-        _user.Number = 2.11d;
+        _user.Number = 2.11f;
         var hash1 = _sha1(_user).ToLowerHexString();
 
         Assert.NotEqual(hash0, hash1);
@@ -41,6 +41,6 @@ public class DoubleTests
 
     private class ThisUser : User
     {
-        public double Number { get; set; }
+        public float Number { get; set; }
     }
 }
